fetch extra fields
