+++ client/src/definitions/extraField.ts
@@ -1,0 +1,28 @@
+type ExtraFieldBase = {
+  id: string;
+  name: string;
+  title: string;
+  hintText: string;
+};
+
+type TextExtraField = {
+  type: "TEXT";
+  data: Record<string, never>;
+} & ExtraFieldBase;
+
+type BoolExtraField = {
+  type: "BOOL";
+  data: {
+    trueValue: string;
+    falseValue: string;
+  };
+} & ExtraFieldBase;
+
+type EnumExtraField = {
+  type: "ENUM";
+  data: {
+    values: string[];
+  };
+} & ExtraFieldBase;
+
+export type ExtraField = TextExtraField | BoolExtraField | EnumExtraField;
