+++ client/src/lib/repositories/datasetFilters.ts
@@ -6,13 +6,20 @@ import { toFiltersInfo } from "../transformers/datasetFilters";
 type GetDatasetFiltersInfo = (opts: {
   fetch: Fetch;
   apiToken: string;
+  organizationSiret?: string | null;
 }) => Promise<DatasetFiltersInfo>;
 
 export const getDatasetFiltersInfo: GetDatasetFiltersInfo = async ({
   fetch,
   apiToken,
+  organizationSiret,
 }) => {
-  const url = `${getApiUrl()}/datasets/filters/`;
+  const urlBase = `${getApiUrl()}/datasets/filters`;
+
+  const url = organizationSiret
+    ? `${urlBase}?organization_siret=${organizationSiret}`
+    : `${urlBase}/`;
+
 
   const request = new Request(url, {
     headers: new Headers(getHeaders(apiToken)),
